+++ xcube_geodb/sql/geodb.sql
@@ -1454,6 +1454,65 @@ BEGIN
 END
 $BODY$;
 
+CREATE OR REPLACE FUNCTION public.geodb_show_indexes(collection text)
+    RETURNS TABLE(indexname name)
+    LANGUAGE 'plpgsql'
+AS
+$BODY$
+BEGIN
+    RETURN QUERY EXECUTE format('SELECT indexname FROM pg_indexes WHERE tablename = ''%s''', collection);
+END
+$BODY$;
+
+CREATE OR REPLACE FUNCTION public.geodb_create_index(collection text, property text)
+    RETURNS void
+    LANGUAGE 'plpgsql'
+AS
+$BODY$
+DECLARE
+    idx_name text;
+BEGIN
+    idx_name := geodb_get_index_name(collection, property);
+    IF property = 'geometry' then
+        EXECUTE format('CREATE INDEX %I ON %I USING GIST(%I)', idx_name, collection, property);
+    ELSE
+        EXECUTE format('CREATE INDEX %I ON %I (%I)', idx_name, collection, property);
+    END IF;
+END
+$BODY$;
+
+CREATE OR REPLACE FUNCTION public.geodb_drop_index(collection text, property text)
+    RETURNS void
+    LANGUAGE 'plpgsql'
+AS
+$BODY$
+DECLARE
+    idx_name text;
+BEGIN
+    idx_name := geodb_get_index_name(collection, property);
+    EXECUTE format('DROP INDEX %I', idx_name);
+END
+$BODY$;
+
+CREATE OR REPLACE FUNCTION public.geodb_get_index_name(collection text, property text)
+    RETURNS text
+    LANGUAGE 'plpgsql'
+AS
+$BODY$
+DECLARE
+    idx_name text;
+    collection_shortened text;
+BEGIN
+    idx_name := format('idx_%s_%s', property, collection);
+    collection_shortened := collection;
+    WHILE LENGTH(idx_name) > 63 LOOP
+        collection_shortened := SUBSTR(collection_shortened, 2, LENGTH(collection_shortened));
+        idx_name := format('idx_%s_%s', property, collection_shortened);
+    END LOOP;
+    RETURN idx_name;
+END
+$BODY$;
+
 -- group functions
 
 CREATE OR REPLACE FUNCTION public.geodb_create_role(user_name text, user_group text)